+++ crates/module-system/module-implementations/integration-tests/src/chain_state/mod.rs
@@ -3,8 +3,9 @@ use std::convert::Infallible;
 use sov_chain_state::{BlockGasInfo, ChainState, StateTransition, TransitionInProgress};
 use sov_mock_da::{MockDaSpec, MockHash, MockValidityCond};
 use sov_modules_api::transaction::SequencerReward;
-use sov_modules_api::{Batch, Gas, GasArray, KernelWorkingSet, Spec, StateCheckpoint};
-use sov_sequencer_registry::BatchSequencerOutcome;
+use sov_modules_api::{
+    Batch, BatchSequencerOutcome, Gas, GasArray, KernelWorkingSet, Spec, StateCheckpoint,
+};
 use sov_test_utils::auth::TestAuth;
 use sov_test_utils::generators::value_setter::ValueSetterMessages;
 use sov_test_utils::runtime::optimistic::TestRuntime;